+++ CHANGELOG.md
@@ -1,9 +1,22 @@
+# IndexedSearchAutocomplete Version 1.0.4
+
+## FEATURE
+- If the search-suggestion-div is opened now you can close it by clicking somewhere else
+
+
+
+# IndexedSearchAutocomplete Version 1.0.4
+
+## FEATURE
+- Added a class to the results-div (".search-autocomplete-results") that shows if there are results ("results", "no-results").
+
+
+
 # IndexedSearchAutocomplete Version 1.0.3
 
 ## FEATURE
 - Added Composer Support
 - Edit Readme with "Use MySQL specific fulltext search"
-
 
 
 # IndexedSearchAutocomplete Version 1.0.2